+++ backend/src/models/mod.rs
@@ -2,3 +2,4 @@ pub mod comments;
 pub mod pix;
 pub mod posts;
 pub mod tags;
+pub mod tweets;