+++ backend/src/state.rs
@@ -23,6 +23,8 @@ impl AppState {
         s3_blogpost_bucket_url: String,
         s3_pix_bucket_name: String,
         s3_pix_bucket_url: String,
+        s3_tweet_bucket_name: String,
+        s3_tweet_bucket_url: String,
     ) -> Self {
         let s3 = S3::new(
             s3_account_id,
@@ -32,6 +34,8 @@ impl AppState {
             s3_blogpost_bucket_url,
             s3_pix_bucket_name,
             s3_pix_bucket_url,
+            s3_tweet_bucket_name,
+            s3_tweet_bucket_url,
         )
         .await;
         Self {